Rework for Saving Creatures to Database
Did a major rework of how much data is stored in initialized creature data and how powerups are handled to allow creature data to easily be stored in the database. Now only the bare essentials about a creature that cannot be determined each time a creature is loaded are stored in the creature data. All other data is added after initializing the creature from it's data as powerups. Added power up groups to replace anatomy aspects like claws and in the future items. Removed attributes (may want to add these back in later)